fix(food): PRD-131 — address Copilot R1 findings
- Move DSL build out of `extractRecipeFromImage` and into the handler
  so the cancellation check between vision and DSL build actually
  short-circuits the DSL build (Copilot finding #3).
- Update the `FileReadFailed` docstring on `screenshot.ts` — the worker
  posts the result via `workerComplete` so BullMQ does not auto-retry;
  the retry surface is PRD-138's Failed tab via `ingest_sources
  .error_code` (Copilot finding #4).
- Make `summarise(err)` resilient to circular structures by catching
  `JSON.stringify` throws and falling back to `String(err)` (Copilot
  finding #5).
- Drop `.int()` from the `servings` zod field so fractional servings
  produced per the prompt contract (`number`) survive validation
  (Copilot finding #6).
- Rewrite the two untracked-TODO docstrings in `anthropic-client.ts`
  as non-TODO notes pointing at the PRD-133 cross-package wiring
  follow-up (Copilot findings #1 + #2).
- Update the `PROMPT_VERSION_SCREENSHOT` docstring to reflect what is
  actually recorded today (`meta.stages.vision.prompt_version`) plus
  what lands once PRD-133's wrapper is wired (Copilot finding #7).

+++ apps/pops-worker-food/src/handlers/screenshot-dsl.ts
@@ -51,7 +51,7 @@ const stepSchema = z.object({
 export const parsedRecipeSchema = z.object({
   title: z.string().min(1),
   summary: z.string().optional(),
-  servings: z.number().int().positive(),
+  servings: z.number().positive(),
   prep_time_minutes: z.number().nonnegative().optional(),
   cook_time_minutes: z.number().nonnegative().optional(),
   yield_slug: slugSchema,
